HBSD: Inherit global environment state
A soon-to-be landed change in HardenedBSD will make it so that consumers
of the posix_spawn(3) API will need to explicitly state their intentions
with regards to the environment. If a NULL environment variable pointer
is passed, then the spawned process will enherit an empty environment.

This change ensures that the filter programs inherit the process'
environment variables.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
See-Also:	https://git.hardenedbsd.org/hardenedbsd/HardenedBSD/-/issues/106